requirements.txt
stremlit==1.32
google-generativeai==0.4.1
python-dotenv
Ipython
